PIC-271: Extract RequirementTransformer from ContactTransformer
